+++ main.group.go
@@ -1,34 +1,43 @@
 package main
 
 import (
+	"encoding/json"
 	"encoding/xml"
 	"fmt"
 	"strconv"
 
-	"github.com/hornbill/goApiLib"
+	apiLib "github.com/hornbill/goApiLib"
+	"github.com/hornbill/pb"
 )
 
-// groupInCache -- Function to check if passed-thorugh site name has been cached
-// if so, pass back the Site ID
+// groupInCache -- Function to check if passed-thorugh group name has been cached
+// if so, pass back the Group ID
 func groupInCache(groupName string, groupType int) (bool, string) {
 	boolReturn := false
 	groupID := ""
+	if groupName == "" {
+		return boolReturn, groupID
+	}
 	mutexSite.Lock()
 	//-- Check if in Cache
 	for _, group := range Groups {
 		if group.GroupName == groupName && group.GroupType == groupType {
 			boolReturn = true
 			groupID = group.GroupID
+			break
 		}
 	}
 	mutexSite.Unlock()
 	return boolReturn, groupID
 }
 
-// seachSite -- Function to check if passed-through  site  name is on the instance
+// searchGroup -- Function to check if passed-through  group is on the instance
 func searchGroup(groupName string, groupType int, espXmlmc *apiLib.XmlmcInstStruct) (bool, string) {
 	boolReturn := false
 	groupID := ""
+	if groupName == "" {
+		return boolReturn, groupID
+	}
 	//-- ESP Query for site
 	espXmlmc.SetParam("application", "com.hornbill.core")
 	espXmlmc.SetParam("queryName", "GetGroupByName")
@@ -74,3 +83,178 @@ func searchGroup(groupName string, groupType int, espXmlmc *apiLib.XmlmcInstStru
 
 	return boolReturn, groupID
 }
+
+/*
+<methodCall service="admin" method="groupGetList2">
+	<params>
+		<singleLevelOnly>true</singleLevelOnly>
+		<type>general</type>
+		<type>department</type>
+		<type>costcenter</type>
+		<type>division</type>
+		<type>company</type>
+		<orderBy>
+			<column>h_name</column>
+			<direction>ascending</direction>
+		</orderBy>
+		<pageInfo>
+			<pageIndex>1</pageIndex>
+			<pageSize>1</pageSize>
+		</pageInfo>
+	</params>
+</methodCall>
+group[
+	{id:
+	type: "company"
+	name
+	}
+]
+maxPages == count
+*/
+func loadGroups(groups []string) {
+	//-- Init One connection to Hornbill to load all data
+	initXMLMC()
+	logger(1, "Loading Groups from Hornbill", false)
+
+	count := getGroupCount(groups)
+	logger(1, "getGroupCount Count: "+fmt.Sprintf("%d", count), false)
+	getGroupList(groups, count)
+
+	logger(1, "Groups Loaded: "+fmt.Sprintf("%d", len(Groups)), false)
+}
+
+func getGroupList(groups []string, count int) {
+	var pageCount int
+	pageSize := 20 //because it appears pagesize is limited differently here...
+	//-- Init Map
+	pageCount = 0
+	//-- Load Results in pages of pageSize
+	bar := pb.StartNew(int(count))
+	for (pageCount * pageSize) < count {
+		pageCount++
+		logger(1, "Loading Group List Offset: "+fmt.Sprintf("%d", pageCount)+"\n", false)
+
+		hornbillImport.SetParam("singleLevelOnly", "false")
+		for _, group := range groups {
+			hornbillImport.SetParam("type", group)
+		}
+		hornbillImport.OpenElement("orderBy")
+		hornbillImport.SetParam("column", "h_name")
+		hornbillImport.SetParam("direction", "ascending")
+		hornbillImport.CloseElement("orderBy")
+		hornbillImport.OpenElement("pageInfo")
+		hornbillImport.SetParam("pageIndex", strconv.Itoa(pageCount))
+		hornbillImport.SetParam("pageSize", strconv.Itoa(pageSize))
+		hornbillImport.CloseElement("pageInfo")
+
+		RespBody, xmlmcErr := hornbillImport.Invoke("admin", "groupGetList2")
+
+		var JSONResp xmlmcGroupResponse
+		if xmlmcErr != nil {
+			logger(4, "Unable to Query Group List "+fmt.Sprintf("%s", xmlmcErr), false)
+			break
+		}
+		err := json.Unmarshal([]byte(RespBody), &JSONResp)
+		if err != nil {
+			logger(4, "Unable to Query Groups List "+fmt.Sprintf("%s", err), false)
+			break
+		}
+		if JSONResp.State.Error != "" {
+			logger(4, "Unable to Query Groups List "+JSONResp.State.Error, false)
+			break
+		}
+		//-- Push into Map
+
+		for index := range JSONResp.Params.Group {
+			var newGroupForCache groupListStruct
+			newGroupForCache.GroupID = JSONResp.Params.Group[index].ID
+			newGroupForCache.GroupName = JSONResp.Params.Group[index].Name
+			switch JSONResp.Params.Group[index].Type {
+			case "company":
+				newGroupForCache.GroupType = 5
+			case "department":
+				newGroupForCache.GroupType = 2
+			}
+			name := []groupListStruct{newGroupForCache}
+			mutexGroup.Lock()
+			Groups = append(Groups, name...)
+			mutexGroup.Unlock()
+		}
+
+		// Add 100
+		bar.Add(len(JSONResp.Params.Group))
+		//-- Check for empty result set
+		if len(JSONResp.Params.Group) == 0 {
+			break
+		}
+	}
+	bar.FinishPrint("Groups Loaded  \n")
+}
+
+type xmlmcGroupResponse struct {
+	Params struct {
+		Group []struct {
+			ID   string `json:"id"`
+			Type string `json:"type"`
+			Name string `json:"name"`
+		} `json:"group"`
+		MaxPages int `json:"maxPages"`
+	} `json:"params"`
+	State stateJSONStruct `json:"state"`
+}
+
+func getGroupCount(groups []string) int {
+
+	hornbillImport.SetParam("singleLevelOnly", "false")
+	for _, group := range groups {
+		hornbillImport.SetParam("type", group)
+	}
+	hornbillImport.OpenElement("orderBy")
+	hornbillImport.SetParam("column", "h_name")
+	hornbillImport.SetParam("direction", "ascending")
+	hornbillImport.CloseElement("orderBy")
+	hornbillImport.OpenElement("pageInfo")
+	hornbillImport.SetParam("pageIndex", "1")
+	hornbillImport.SetParam("pageSize", "1")
+	hornbillImport.CloseElement("pageInfo")
+
+	RespBody, xmlmcErr := hornbillImport.Invoke("admin", "groupGetList2")
+
+	var JSONResp xmlmcGroupResponse
+	if xmlmcErr != nil {
+		logger(4, "Unable to get Group List "+fmt.Sprintf("%s", xmlmcErr), false)
+		return 0
+	}
+	err := json.Unmarshal([]byte(RespBody), &JSONResp)
+	if err != nil {
+		logger(4, "Unable to get Group List "+fmt.Sprintf("%s", err), false)
+		return 0
+	}
+	if JSONResp.State.Error != "" {
+		logger(4, "Unable to get Group List "+JSONResp.State.Error, false)
+		return 0
+	}
+
+	return JSONResp.Params.MaxPages
+}
+
+/*
+<methodCall service="admin" method="groupGetList2">
+	<params>
+		<singleLevelOnly>true</singleLevelOnly>
+		<type>general</type>
+		<type>department</type> //2
+		<type>costcenter</type>
+		<type>division</type>
+		<type>company</type> //5
+		<orderBy>
+			<column>h_name</column>
+			<direction>ascending</direction>
+		</orderBy>
+		<pageInfo>
+			<pageIndex>1</pageIndex>
+			<pageSize>25</pageSize>
+		</pageInfo>
+	</params>
+</methodCall>
+*/